% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFEED_core_functions.R, R/imports.R
\docType{package}
\name{UFEED-package}
\alias{UFEED-package}
\title{UFEED package imports}
\description{
Downloads weather and soil data and computes environmental features for plant physiology modeling.

Downloads weather and soil data and computes environmental features for plant physiology modeling.
}
\author{
\strong{Maintainer}: Hongrui Wang \email{hongrui.wang@inrae.fr}

}
\keyword{internal}
