% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{UFEED-package}
\alias{UFEED}
\alias{UFEED-package}
\title{Package imports}
\description{
Downloads weather and soil data and computes environmental feature sets for grapevine and crop modeling applications.
}
\author{
\strong{Maintainer}: Hongrui Wang \email{hongrui.wang@inrae.fr}

Authors:
\itemize{
  \item Hongrui Wang \email{hongrui.wang@inrae.fr}
}

}
\keyword{internal}
